+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 Changed
   * AZ-398: Force lowercases on default generated name
 
+Added
+  * AZ-365: Add diagnostics settings
+
+Updated
+  * AZ-273: Module now compatible terraform `v0.13+`
+
 # v3.1.0 - 2020-08-25
 
 Breaking